reviewersPayload small refactoring
