+++ dxcherry/src/reviewers/reviewer.ts
@@ -1,0 +1,11 @@
+interface Reviewer {
+    l: string,
+    f: string,
+    gh: string,
+    e: string,
+    t: string
+};
+
+export {
+    Reviewer
+};
